pyproject.toml
[project]
name = "vector-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "jupyter>=1.1.1",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.14",
    "langchain-ollama>=0.2.2",
    "langchain-postgres>=0.0.12",
    "langgraph>=0.2.61",
    "libmagic>=1.0",
    "numpy>=1.26.4",
    "psycopg2>=2.9.10",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "unstructured>=0.16.12",
    "urllib3>=2.3.0",
]
